Use bitmap intrinsic dimensions instead of view dimensions
When the shimmer layout only has a single child that's an image, we can significantly reduce the memory footprint for bitmaps we generate by using the intrinsic height and width of the ImageView's bitmap. In this case we'll end up generating masks that match the size of the bitmap we're rendering, instead of generating bitmaps that match the entire size of the view.