Rewrite ShimmerFrameLayout to be much more performant
Summary:
A rewrite of shimmer to take into account memory/perf issues

The previous version of shimmer would require 3 extra bitmaps to achieve its visual effect.
This version only requires 1 extra hardware layer for views that contain transparency and 0 for those that do
not. This significantly reduces memory consumption in both the Java world and GPU world. It also increases performance.

It does this by taking advantage of the ability compose the effect via a Paint + hardware layer + Porter Duff xfer modes.
In the case of fully opaque views, the shimmer is simply drawn on top of the content.

A preliminary check of the effects of this implementation on the sample app show a 15MB reduction in Java heap usage.

Reviewed By: sjkirby

Differential Revision: D7001991

fbshipit-source-id: 9e50a34afee0747b4ca9ba7cf96d5576d5e671c0