Fix ShimmerDrawable rect calculation when using InsetDrawable
Summary:
`ShimmerDrawable` does not work too well when used as the inner drawable of an `InsetDrawable`.

Consider the following code

```
InsetDrawable wrapShimmerDrawableWithInset(ShimmerDrawable shimmerDrawable) {
  return new InsetDrawable(shimmerDrawable, 10)
}
```

Instead of producing a drawable that is inset by 10px on all sides, this will instead produce a drawable visually that:

- Has no inset on the top or left
- Has an inset of 20 on the bottom and right

This is because we're not taking into account the starting position of the bounds passed to `ShimmerDrawable` in `onBoundsChange()`, assuming that we're always starting at `0, 0` (which is not the case when used inside an `InsetDrawable`.

Fixing up by instead obeying the bounds given to through `ShimmerDrawable#onBoundsChanged()` when calculating our drawing rect.

Reviewed By: xiphirx

Differential Revision: D22869773

fbshipit-source-id: d7bcfcc7f4f6a09e25b1cab4355222c75daea8e7